Introduce `KVStore` trait and `FilesystemStore` impl
Rather than further relying on the upstream `KVStorePersister`, we here
implement a general `KVStore` trait that allows access to
`Read`s/`TransactionalWrite`s which may be used to deserialize/serialize
data via the `Readable`/`Writeable` implementations.

Notably `TransactionalWrite` is a `Write` for which the written data
needs to be explictly `commit`ed, asserting that we always persist
either the whole new change or no change at all.