Introduce `OnchainFeeEstimator`
We also decouple fee estimation from the BDK on-chain wallet and BDK's
corresponding `EsploraBlockchain`. Instead we use the esplora client
directly in a dedicated `OnchainFeeEstimator` object.

For one, this change is nice as it allows us
to move more things out of the `Wallet` thread and corresponding locks.
Moreover, it makes things more modular in general, which makes future
upgrades and testing easier.